fix(backends): ensure that returned date results are actually proper date values
BREAKING CHANGE: Columns with Ibis `date` types are now returned as object dtype containing `datetime.date` objects when executing with the pandas backend.

+++ ibis/backends/polars/tests/test_udf.py
@@ -46,7 +46,7 @@ def test_multiple_argument_udf(alltypes):
     result = expr.execute()
 
     df = alltypes[["smallint_col", "int_col"]].execute()
-    expected = (df.smallint_col + df.int_col).astype("int32")
+    expected = df.smallint_col + df.int_col
 
     tm.assert_series_equal(result, expected.rename("tmp"))
